+++ R/likelihood.R
@@ -432,8 +432,10 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
   # add mu.center back to the mean value after kalman filter / mean profiling
   # pre-standardizing the data should also help
   SCALE <- sqrt(mean(get.telemetry(data,axes)^2))
-  data <- unit.telemetry(data,length=SCALE)
-  CTMM <- unit.ctmm(CTMM,length=SCALE)
+  # standardize time by median diff time
+  TSCALE <- stats::median(diff(data$t))
+  data <- unit.telemetry(data,length=SCALE,time=TSCALE)
+  CTMM <- unit.ctmm(CTMM,length=SCALE,time=TSCALE)
 
   # used for minimum scale of parameter inspection
   n <- length(data$t)
@@ -445,7 +447,7 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
   # unstandardize (includes likelihood adjustment)
   unscale.ctmm <- function(CTMM)
   {
-    CTMM <- unit.ctmm(CTMM,length=1/SCALE)
+    CTMM <- unit.ctmm(CTMM,length=1/SCALE,time=1/TSCALE)
     # log-likelihood adjustment
     CTMM$loglike <- CTMM$loglike - length(axes)*n*log(SCALE)
     # translate back to origin from center
@@ -486,19 +488,20 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
   CTMM <- ctmm.prepare(data,CTMM,tau=FALSE) # don't muck with taus
   UERE <- attr(CTMM$error.mat,"flag") # do we fit the error? Need to know for optimization
 
-  # id and characterize parameters for profiling
+  ### id and characterize parameters for profiling ###
   pars <- NAMES <- parscale <- lower <- upper <- period <- NULL
   ORIGINAL <- CTMM # original structure of model before fitting
+  linear.cov <- FALSE # represent sigma linearly (for perturbation) versus * (for optimization)
   setup.parameters <- function(CTMM,profile=TRUE,linear=FALSE)
   {
-    STUFF <- id.parameters(CTMM,profile=profile,linear=linear,UERE=UERE,dt=dt,df=df,dz=dz,STRUCT=ORIGINAL)
+    STUFF <- id.parameters(CTMM,profile=profile,linear=linear,linear.cov=linear.cov,UERE=UERE,dt=dt,df=df,dz=dz,STRUCT=ORIGINAL)
     NAMES <<- STUFF$NAMES
     parscale <<- STUFF$parscale
     lower <<- STUFF$lower
     upper <<- STUFF$upper
     period <<- STUFF$period
     # initial guess for optimization
-    pars <<- get.parameters(CTMM,NAMES)
+    pars <<- get.parameters(CTMM,NAMES,linear.cov=linear.cov)
   }
   setup.parameters(CTMM)
   if("error" %nin% NAMES) { CTMM$error <- as.logical(CTMM$error) } # fix numeric error when it should be logical
@@ -511,8 +514,8 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
   fn <- function(p,zero=0)
   {
     names(p) <- NAMES
-    p <- clean.parameters(p)
-    CTMM <- set.parameters(CTMM,p)
+    p <- clean.parameters(p,linear.cov=linear.cov)
+    CTMM <- set.parameters(CTMM,p,linear.cov=linear.cov)
 
     # negative log likelihood
     return(-ctmm.loglike(data,CTMM,REML=REML,zero=-zero,profile=profile))
@@ -532,7 +535,7 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
     return(COV)
   }
 
-  # NOW OPTIMIZE
+  ### NOW OPTIMIZE ###
   profile <- TRUE
   if(length(NAMES)==0) # EXACT
   {
@@ -556,7 +559,7 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
     COVSTUFF <- COV.covm(CTMM$sigma,n=n,k=k.mean,REML=REML)
     CTMM$COV <- COVSTUFF$COV
   }
-  else # all further cases require optimization
+  else ### all further cases require optimization ###
   {
     if(trace) { message("Maximizing likelihood.") }
     control$covariance <- covariance()
@@ -570,9 +573,9 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
     store.pars <- function(pars,profile=TRUE,finish=TRUE)
     {
       names(pars) <- NAMES
-      pars <- clean.parameters(pars)
+      pars <- clean.parameters(pars,linear.cov=linear.cov)
 
-      CTMM <<- set.parameters(CTMM,pars)
+      CTMM <<- set.parameters(CTMM,pars,linear.cov=linear.cov)
 
       # this is a wasted evaluation !!! store verbose glob in environment?
       if(finish) { CTMM <<- ctmm.loglike(data,CTMM,REML=REML,verbose=TRUE,profile=profile) }
@@ -580,10 +583,10 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
     store.pars(pars,finish=TRUE)
 
     profile <- FALSE # no longer solving covariance analytically
+    if(method %in% c("pREML","pHREML")) { linear.cov <- TRUE }
     setup.parameters(CTMM,profile=FALSE)
-
     ### COV CALCULATION #############
-    if(COV || method %in% c("pREML","pHREML","HREML"))
+    if(COV || method %in% c("pREML","pHREML"))
     {
       if(trace) { message("Calculating Hessian.") }
       DIFF <- genD(par=pars,fn=fn,zero=-CTMM$loglike,lower=lower,upper=upper,parscale=parscale,Richardson=2,mc.cores=1)
@@ -604,21 +607,15 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
     else
     { empty.env(MLE.env) }
 
-    # pREML correction ############################### only do pREML if sufficiently away from boundaries
+    ### pREML correction ########## only do pREML if sufficiently away from boundaries
     if(method %in% c("pREML","pHREML") && mat.min(hess) > .Machine$double.eps*length(NAMES))
     {
       # parameter correction
       REML <- TRUE
       #ML.grad <- grad # save old ML gradient
       if(trace) { message("Calculating REML gradient.") }
-      DIFF <- genD(par=pars,fn=fn,zero=-CTMM$loglike,lower=lower,upper=upper,parscale=parscale,Richardson=2,order=1,mc.cores=1)
+      DIFF <- genD(par=pars,fn=fn,lower=lower,upper=upper,parscale=parscale,Richardson=2,order=1,mc.cores=1)
 
-      # trying to make this robust here
-      # COV is -1/Hessian, grad is of -loglike
-      # using least-squares solution in the case of degeneracy (under parscale natural units)
-      # hess <- t(hess*parscale)*parscale
-      # d.pars <- -c(PDsolve(t(hess)%*%hess,pseudo=TRUE) %*% t(hess)%*%(DIFF$gradient*parscale)) * parscale
-      # OK, that didn't work so well... sticking with this
       d.pars <- -c(CTMM$COV %*% DIFF$gradient)
 
       # increment transformed parameters
@@ -627,12 +624,9 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
       pars <- line.boxer(d.pars,pars,lower=lower,upper=upper,period=period)
       names(pars) <- NAMES
 
-      # store parameter correction only if a correction was made
-      if(method=="pREML")
-      {
-        profile <- FALSE
-        store.pars(pars,profile=FALSE,finish=TRUE)
-      }
+      # store parameter correction
+      profile <- FALSE
+      store.pars(pars,profile=FALSE,finish=TRUE)
     }
     else if(method %in% c("pREML",'pHREML'))
     {
@@ -640,13 +634,14 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
       if(method=='pREML') { method <- 'ML' }
       else if(method=='pHREML') { method <- 'HREML' }
     }
+    linear.cov <- FALSE
+    ### end pREML correction ###
 
-    # profile REML parameters
+    ### profile linear REML parameters ###
     if(method %in% c('pHREML','HREML'))
     {
       REML <- TRUE
       profile <- TRUE
-      store.pars(pars,profile=TRUE,finish=TRUE)
 
       # profile REML linear parameters numerically if necessary (error || circle)
       setup.parameters(CTMM,profile=TRUE,linear=TRUE)
@@ -659,31 +654,28 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
         control$zero <- TRUE
         RESULT <- optimizer(par=pars,fn=fn,method=op.method,lower=lower,upper=upper,period=period,control=control)
         pars <- clean.parameters(RESULT$par)
-
-        store.pars(pars,profile=TRUE,finish=TRUE)
       }
+      # includes free profile
+      store.pars(pars,profile=TRUE,finish=TRUE)
     }
 
-    # FINAL COVARIANCE ESTIMATE
-    TEST <- method %in% c('pREML','pHREML','HREML')
-    if(TEST && COV) ### CALCULATE COVARIANCE MATRIX ###
+    ### FINAL COVARIANCE ESTIMATE ###
+    if(COV && method %in% c('pREML','pHREML','HREML')) ### CALCULATE COVARIANCE MATRIX ###
     {
       profile <- FALSE
       setup.parameters(CTMM,profile=FALSE)
 
       if(trace) { message("Calculating REML Hessian.") }
       # calcualte REML Hessian at pREML parameters
-      DIFF <- genD(par=pars,fn=fn,zero=-CTMM$loglike,lower=lower,upper=upper,parscale=parscale,Richardson=2,mc.cores=1)
+      DIFF <- genD(par=pars,fn=fn,lower=lower,upper=upper,parscale=parscale,Richardson=2,mc.cores=1)
       # Using MLE gradient, which should be zero off boundary
       CTMM$COV <- cov.loglike(DIFF$hessian,grad)
     }
-    else if(TEST) ### don't confuse the ML COV with pREML COV
-    { CTMM$COV <- NULL }
 
     if(COV) { dimnames(CTMM$COV) <- list(NAMES,NAMES) }
   } # end optimized estimates
 
-  # model likelihood
+  # model likelihood (not REML for AIC)
   if(method!='ML') { CTMM$loglike <- ctmm.loglike(data,CTMM=CTMM,REML=FALSE,profile=FALSE) }
   CTMM$method <- method
 
@@ -696,6 +688,9 @@ ctmm.fit <- function(data,CTMM=ctmm(),method="pHREML",COV=TRUE,control=list(),tr
 
   # calculate AIC,AICc,BIC,MSPE,...
   CTMM <- ic.ctmm(CTMM,n)
+
+  # would be temporary ML COV for pREML/pHREML
+  if(!COV) { CTMM$COV <- NULL }
 
   return(CTMM)
 }
@@ -728,6 +723,9 @@ ic.ctmm <- function(CTMM,n)
   { CTMM$AICc <- -2*CTMM$loglike + (q*n)^2/(q*n+q*k.mean) * 2*k/(q*n-k-nu) }
   else if(method %in% c('pHREML','HREML','REML'))
   { CTMM$AICc <- -2*CTMM$loglike + (q*n-q*k.mean) * 2*k/(q*n-k-nu) }
+
+  # fix divergence
+  if(q*n<=k+nu) { CTMM$AICc <- Inf }
 
   # Mean square prediction error
   mspe <- function(K=1)